Move company to CRM directory
